requirements.txt
entsoe.py==0.7.10
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.4.2
pandas==3.0.0
pytz==2025.2
requests==2.32.5